Update TapTargets for Start Destination
